--- src/shared/utils/tag-name/index.ts
@@ -1,1 +1,0 @@
-export * from './ui';
