--- src/shared/utils/tag-name/types.ts
@@ -1,7 +1,0 @@
-import {ElementType, ReactNode} from 'react';
-
-export type TagNameProps = {
-  as?: ElementType;
-  children: ReactNode;
-  className?: string;
-}
