--- src/shared/utils/tag-name/ui.tsx
@@ -1,6 +1,0 @@
-import {FC} from 'react';
-import {TagNameProps} from './types';
-
-export const TagName: FC<TagNameProps> = ({ as: Tag = 'span', ...otherProps }) => (
-  <Tag {...otherProps} />
-);
