--- src/shared/utils/visually-hidden/index.ts
@@ -1,1 +1,0 @@
-export * from './ui';
