+++ src/shared/utils/visually-hidden/index.tsx
